Fix Kalorienberechnung Int Fehler

+++ app/src/main/java/mi/ur/de/android/runnersmeetup/CalculatorService.java
@@ -77,7 +77,7 @@ public class CalculatorService extends Service implements CalculatorListener {
                 if (lastLocation != null) {
                     currentDistance += location.distanceTo(lastLocation);
                 }
-                kcal = (int) currentDistance/1000 * Constants.getWeight();
+                kcal = (int) (currentDistance/1000.0 * (double) Constants.getWeight());
                 currentVelocity = location.getSpeed() * 3.6;
                 updateVelocityView(currentVelocity);
                 updateDistanceView(currentDistance);
